revamped frame rate calculation

+++ source/overlay[dll]/overlay.h
@@ -2,10 +2,8 @@
 
 extern PUSH_SHARED_MEMORY *PushSharedMemory;
 extern BOOLEAN g_DXGI;
-//extern UINT8    PushRefreshRate;
-//extern UINT8 AcceptableFps;
 extern UINT16 DiskResponseTime;
-extern UINT32 FrameRate;
+extern double FrameRate;
 extern BOOLEAN IsStableFramerate;
 
 char *GetDirectoryFile(char *pszFileName);
